Refactor #89 - Extract test help method in `testkit` module
