+++ app/src/test/java/com/nlab/reminder/core/state/StateMachineHandleScopeTest.kt
@@ -17,6 +17,7 @@
 package com.nlab.reminder.core.state
 
 import com.nlab.reminder.test.*
+import com.nlab.testkit.genInt
 import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.test.runTest
