+++ app/src/test/java/com/nlab/reminder/domain/common/schedule/impl/DefaultScheduleUiStateFlowFactoryTest.kt
@@ -17,7 +17,7 @@
 package com.nlab.reminder.domain.common.schedule.impl
 
 import com.nlab.reminder.domain.common.schedule.*
-import com.nlab.reminder.test.genBoolean
+import com.nlab.testkit.genBoolean
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.test.runTest
