+++ app/src/test/java/com/nlab/reminder/domain/common/schedule/impl/DefaultUpdateCompleteUseCaseTest.kt
@@ -23,6 +23,10 @@ import com.nlab.reminder.domain.common.util.link.transaction.genTransactionIdGen
 import com.nlab.reminder.domain.common.schedule.*
 import com.nlab.reminder.domain.common.util.link.transaction.genTransactionId
 import com.nlab.reminder.test.*
+import com.nlab.testkit.genBoolean
+import com.nlab.testkit.genBothify
+import com.nlab.testkit.genLong
+import com.nlab.testkit.once
 import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.test.advanceTimeBy
