+++ app/src/test/java/com/nlab/reminder/domain/feature/schedule/all/AllScheduleViewModelTest.kt
@@ -22,7 +22,7 @@ import com.nlab.reminder.core.state.StateMachine
 import com.nlab.reminder.core.state.asContainer
 import com.nlab.reminder.domain.feature.home.*
 import com.nlab.reminder.test.genFlowObserveCoroutineScope
-import com.nlab.reminder.test.once
+import com.nlab.testkit.once
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
